+++ services/web/server/src/simcore_service_webserver/storage/_rest.py
@@ -7,20 +7,14 @@ import logging
 import urllib.parse
 from typing import Any, Final, NamedTuple
 from urllib.parse import quote, unquote
-from uuid import UUID
 
 from aiohttp import ClientTimeout, web
-from models_library.api_schemas_long_running_tasks.base import TaskProgress
 from models_library.api_schemas_long_running_tasks.tasks import (
     TaskGet,
-    TaskResult,
-    TaskStatus,
 )
 from models_library.api_schemas_rpc_async_jobs.async_jobs import (
-    AsyncJobId,
     AsyncJobNameData,
 )
-from models_library.api_schemas_storage import STORAGE_RPC_NAMESPACE
 from models_library.api_schemas_storage.storage_schemas import (
     FileUploadCompleteResponse,
     FileUploadCompletionBody,
@@ -44,12 +38,6 @@ from servicelib.aiohttp.requests_validation import (
 )
 from servicelib.aiohttp.rest_responses import create_data_response
 from servicelib.common_headers import X_FORWARDED_PROTO
-from servicelib.rabbitmq.rpc_interfaces.async_jobs.async_jobs import (
-    abort,
-    get_result,
-    get_status,
-    list_jobs,
-)
 from servicelib.rabbitmq.rpc_interfaces.storage.data_export import start_data_export
 from servicelib.rabbitmq.rpc_interfaces.storage.paths import (
     compute_path_size as remote_compute_path_size,
@@ -63,7 +51,7 @@ from ..login.decorators import login_required
 from ..models import RequestContext
 from ..rabbitmq import get_rabbitmq_rpc_client
 from ..security.decorators import permission_required
-from ._exception_handlers import handle_data_export_exceptions
+from ..tasks._exception_handlers import handle_data_export_exceptions
 from .schemas import StorageFileIDStr
 from .settings import StorageSettings, get_plugin_settings
 
@@ -477,135 +465,4 @@ async def export_data(request: web.Request) -> web.Response:
             result_href=f"{request.url.with_path(str(request.app.router['get_async_job_result'].url_for(job_id=_job_id)))}",
         ),
         status=status.HTTP_202_ACCEPTED,
-    )
-
-
-@routes.get(
-    _storage_prefix + "/async-jobs",
-    name="get_async_jobs",
-)
-@login_required
-@permission_required("storage.files.*")
-@handle_data_export_exceptions
-async def get_async_jobs(request: web.Request) -> web.Response:
-    _req_ctx = RequestContext.model_validate(request)
-
-    rabbitmq_rpc_client = get_rabbitmq_rpc_client(request.app)
-
-    user_async_jobs = await list_jobs(
-        rabbitmq_rpc_client=rabbitmq_rpc_client,
-        rpc_namespace=STORAGE_RPC_NAMESPACE,
-        job_id_data=AsyncJobNameData(
-            user_id=_req_ctx.user_id, product_name=_req_ctx.product_name
-        ),
-        filter_="",
-    )
-    return create_data_response(
-        [
-            TaskGet(
-                task_id=f"{job.job_id}",
-                task_name=f"{job.job_id}",
-                status_href=f"{request.url.with_path(str(request.app.router['get_async_job_status'].url_for(job_id=str(job.job_id))))}",
-                abort_href=f"{request.url.with_path(str(request.app.router['abort_async_job'].url_for(job_id=str(job.job_id))))}",
-                result_href=f"{request.url.with_path(str(request.app.router['get_async_job_result'].url_for(job_id=str(job.job_id))))}",
-            )
-            for job in user_async_jobs
-        ],
-        status=status.HTTP_200_OK,
-    )
-
-
-class _StorageAsyncJobId(BaseModel):
-    job_id: AsyncJobId
-
-
-@routes.get(
-    _storage_prefix + "/async-jobs/{job_id}/status",
-    name="get_async_job_status",
-)
-@login_required
-@permission_required("storage.files.*")
-@handle_data_export_exceptions
-async def get_async_job_status(request: web.Request) -> web.Response:
-    class _PathParams(BaseModel):
-        job_id: UUID
-
-    _req_ctx = RequestContext.model_validate(request)
-    rabbitmq_rpc_client = get_rabbitmq_rpc_client(request.app)
-
-    async_job_get = parse_request_path_parameters_as(_StorageAsyncJobId, request)
-    async_job_rpc_status = await get_status(
-        rabbitmq_rpc_client=rabbitmq_rpc_client,
-        rpc_namespace=STORAGE_RPC_NAMESPACE,
-        job_id=async_job_get.job_id,
-        job_id_data=AsyncJobNameData(
-            user_id=_req_ctx.user_id, product_name=_req_ctx.product_name
-        ),
-    )
-    _job_id = f"{async_job_rpc_status.job_id}"
-    return create_data_response(
-        TaskStatus(
-            task_progress=TaskProgress(
-                task_id=_job_id, percent=async_job_rpc_status.progress.actual_value
-            ),
-            done=async_job_rpc_status.done,
-            started=None,
-        ),
-        status=status.HTTP_200_OK,
-    )
-
-
-@routes.post(
-    _storage_prefix + "/async-jobs/{job_id}:abort",
-    name="abort_async_job",
-)
-@login_required
-@permission_required("storage.files.*")
-@handle_data_export_exceptions
-async def abort_async_job(request: web.Request) -> web.Response:
-    class _PathParams(BaseModel):
-        job_id: UUID
-
-    _req_ctx = RequestContext.model_validate(request)
-
-    rabbitmq_rpc_client = get_rabbitmq_rpc_client(request.app)
-    async_job_get = parse_request_path_parameters_as(_StorageAsyncJobId, request)
-    await abort(
-        rabbitmq_rpc_client=rabbitmq_rpc_client,
-        rpc_namespace=STORAGE_RPC_NAMESPACE,
-        job_id=async_job_get.job_id,
-        job_id_data=AsyncJobNameData(
-            user_id=_req_ctx.user_id, product_name=_req_ctx.product_name
-        ),
-    )
-    return web.Response(status=status.HTTP_204_NO_CONTENT)
-
-
-@routes.get(
-    _storage_prefix + "/async-jobs/{job_id}/result",
-    name="get_async_job_result",
-)
-@login_required
-@permission_required("storage.files.*")
-@handle_data_export_exceptions
-async def get_async_job_result(request: web.Request) -> web.Response:
-    class _PathParams(BaseModel):
-        job_id: UUID
-
-    _req_ctx = RequestContext.model_validate(request)
-
-    rabbitmq_rpc_client = get_rabbitmq_rpc_client(request.app)
-    async_job_get = parse_request_path_parameters_as(_PathParams, request)
-    async_job_rpc_result = await get_result(
-        rabbitmq_rpc_client=rabbitmq_rpc_client,
-        rpc_namespace=STORAGE_RPC_NAMESPACE,
-        job_id=async_job_get.job_id,
-        job_id_data=AsyncJobNameData(
-            user_id=_req_ctx.user_id, product_name=_req_ctx.product_name
-        ),
-    )
-
-    return create_data_response(
-        TaskResult(result=async_job_rpc_result.result, error=None),
-        status=status.HTTP_200_OK,
     )
